AssignMate[Backend]/pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "3.12.6"
django = "6.0.2"
djangorestframework = "3.16.1"
django-debug-toolbar = "6.2.0"
drf-nested-routers = "0.95.0"
django-safedelete = "1.4.1"
django-cors-headers = "4.9.0"
djangorestframework-simplejwt = "5.5.1"
django-environ = "0.13.0"
psycopg = {extras = ["binary"], version = "3.3.3"}
gunicorn = "25.1.0"
